requirements.txt
click>=8.1.8
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.23.2
openai>=1.68.2
python-dotenv==1.0.1

# Required by crewai-tools

tiktoken >= 0.8.0
langsmith>=0.3.18,<0.4.0
python-multipart

# Data and ML
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
jinja2==3.1.4
jsonschema==4.22.0
onnxruntime==1.22.0
oauthlib==3.2.2

# Observability
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-proto>=1.30.0
opentelemetry-exporter-otlp-proto-common>=1.30.0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-instrumentation>=0.47b0
opentelemetry-instrumentation-asgi>=0.47b0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-semantic-conventions>=0.47b0
opentelemetry-util-http>=0.47b0

pydantic==2.11.6
protobuf>=5.0.0,<6.0.0
